+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/KlumValidationException.java
@@ -23,51 +23,38 @@
  */
 package com.blackbuild.klum.ast.util;
 
-import com.blackbuild.klum.ast.util.layer3.KlumVisitorException;
+import com.blackbuild.groovy.configdsl.transform.Validate;
 
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 
+import static java.util.stream.Collectors.joining;
+
+/**
+ * Denotes an exception that means a validation error occurred in the model.
+ * This is used to collect multiple validation errors and provide a detailed message.
+ * <p>
+ *     Note that this is explicitly not a subclass of {@link KlumModelException}, since a KlumModelException
+ *     points to a specific model element that is invalid, while this exception collects multiple validation errors.
+ * </p>
+ */
 public class KlumValidationException extends KlumException {
 
-    private final Map<String, List<KlumVisitorException>> validationErrors = new LinkedHashMap<>();
+    private final List<KlumValidationResult> validationResults;
 
-    public KlumValidationException(List<KlumVisitorException> validationErrors) {
-        for (KlumVisitorException validationError : validationErrors) {
-            addException(validationError);
-        }
-    }
-
-    private void addException(KlumVisitorException validationError) {
-        addSuppressed(validationError);
-        this.validationErrors.computeIfAbsent(validationError.getBreadcrumbPath(), k -> new ArrayList<>()).add(validationError);
-    }
-
-    public KlumValidationException() {
-    }
-
-    public void merge(KlumValidationException other) {
-        for (List<KlumVisitorException> errors : other.getValidationErrors().values()) {
-            for (KlumVisitorException validationError : errors) {
-                addException(validationError);
-            }
-        }
-    }
-
-    public Map<String, List<KlumVisitorException>> getValidationErrors() {
-        return validationErrors;
+    public KlumValidationException(List<KlumValidationResult> validationResults) {
+        this.validationResults = validationResults;
     }
 
     @Override
     public String getMessage() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("Validation errors:\n");
-        validationErrors.forEach((key, value) -> {
-            sb.append("  at ").append(key).append(":\n");
-            value.forEach(e -> sb.append("    - ").append(e.getUnlocalizedMessage()).append("\n"));
-        });
-        return sb.toString();
+        return validationResults.stream()
+                .map(KlumValidationResult::getMessage)
+                .collect(joining("\n"));
+    }
+
+    public String getMessage(Validate.Level minimumLevel) {
+        return validationResults.stream()
+                .map(klumValidationResult -> klumValidationResult.getMessage(minimumLevel))
+                .collect(joining("\n"));
     }
 }